# data/swagger.json
{
   "swagger": "2.0",
   "info": {
      "description": "API for Fixthis",
      "version": "0.1.0",
      "title": "Fixthis",
      "termsOfService": ""
   },
   "basePath": "/api",
...
   "schemes": [
      "http"
   ],
   "paths": {
...
      "/profile/removeportfolio": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "Remove Portfolio",
            "description": "Api to remove user portfolio",
            "operationId": "user.removeportfolio",
            "consumes": [
               "multipart/form-data"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "portfolioid",
                  "type": "integer",
                  "required": true,
                  "description": "portfolio id to remove"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/profile/handlegeneralinfo": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "Handle general user info in more tab",
            "description": "Api to handle general user info in more tab",
            "operationId": "user.handlegeneralinfo",
            "consumes": [
               "multipart/form-data"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "first_name",
                  "type": "string",
                  "description": "First name"
               },
               {
                  "in": "formData",
                  "name": "last_name",
                  "type": "string",
                  "description": "last name"
               },
               {
                  "in": "formData",
                  "name": "location",
                  "type": "string",
                  "description": "location"
               },
               {
                  "in": "formData",
                  "name": "tagline",
                  "type": "string",
                  "description": "Tag Line"
               },
               {
                  "in": "formData",
                  "name": "about_me",
                  "type": "string",
                  "description": "About me"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
...
      "/user/updatecheckinfo": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "Api to update mandatory info before placing order",
            "description": "",
            "operationId": "user.updatecheckinfo",
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "fieldType",
                  "description": "Type of field to be updated",
                  "type": "string",
                  "required": true,
                  "enum": [
                     "dob",
                     "billing_address"
                  ]
               },
               {
                  "in": "formData",
                  "name": "value",
                  "description": "Value to be set.In case of DOB it should be yyyy-mm-dd \n for billing address {\"city\":\"North Miami Beach\",\"state\":\"Florida\",\"street_address\":\"your address here\",\"zipcode\":\"2222\"}",
                  "type": "string",
                  "required": true
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
...
   },
...
   "securityDefinitions": {
      "bearer": {
         "in": "header",
         "name": "Authorization",
         "type": "apiKey"
      }
   }
}